Changed the variable names for the model classes

+++ src/main/java/com/example/mescfit/exercise/ExerciseRepository.java
@@ -1,10 +1,9 @@
 package com.example.mescfit.exercise;
 
-import com.example.mescfit.model.Exercise;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExerciseRepository extends JpaRepository<Exercise, Integer> {
+public interface ExerciseRepository extends JpaRepository<Exercise, Long> {
 
 }
