+++ src/main/java/com/example/mescfit/exercise/ExerciseService.java
@@ -1,7 +1,6 @@
 package com.example.mescfit.exercise;
 
 import com.example.mescfit.exceptions.NotFoundException;
-import com.example.mescfit.model.Exercise;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -21,7 +20,7 @@ public class ExerciseService {
         return this.exerciseRepository.save(exercise);
     }
 
-    public Exercise getExercise(Integer id) {
+    public Exercise getExercise(Long id) {
         return this.exerciseRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException(String.format("Exercise with id %s does was not found", id)));
     }
